Change the DnsMessage.id() to be a short.
Motivation:

The id of a DnsMessage should be a 16 bit identifier accourding to RFC 1035 Section 4.1.1.

See also https://www.ietf.org/rfc/rfc1035.txt

Modifications:

Change id from unsigned short to short to be compliant with the spec.

Result:

Correctly handle dns message id.

+++ codec-dns/src/main/java/io/netty/handler/codec/dns/DatagramDnsQuery.java
@@ -37,7 +37,7 @@ public class DatagramDnsQuery extends DefaultDnsQuery
      * @param id the {@code ID} of the DNS query
      */
     public DatagramDnsQuery(
-            InetSocketAddress sender, InetSocketAddress recipient, int id) {
+            InetSocketAddress sender, InetSocketAddress recipient, short id) {
         this(sender, recipient, id, DnsOpCode.QUERY);
     }
 
@@ -50,7 +50,7 @@ public class DatagramDnsQuery extends DefaultDnsQuery
      * @param opCode the {@code opCode} of the DNS query
      */
     public DatagramDnsQuery(
-            InetSocketAddress sender, InetSocketAddress recipient, int id, DnsOpCode opCode) {
+            InetSocketAddress sender, InetSocketAddress recipient, short id, DnsOpCode opCode) {
         super(id, opCode);
 
         if (recipient == null && sender == null) {
@@ -77,7 +77,7 @@ public class DatagramDnsQuery extends DefaultDnsQuery
     }
 
     @Override
-    public DatagramDnsQuery setId(int id) {
+    public DatagramDnsQuery setId(short id) {
         return (DatagramDnsQuery) super.setId(id);
     }
