+++ codec-dns/src/main/java/io/netty/handler/codec/dns/DnsQuery.java
@@ -20,7 +20,7 @@ package io.netty.handler.codec.dns;
  */
 public interface DnsQuery extends DnsMessage {
     @Override
-    DnsQuery setId(int id);
+    DnsQuery setId(short id);
 
     @Override
     DnsQuery setOpCode(DnsOpCode opCode);
