+++ resolver-dns/src/main/java/io/netty/resolver/dns/DnsQueryContext.java
@@ -45,7 +45,7 @@ final class DnsQueryContext {
 
     private final DnsNameResolver parent;
     private final Promise<AddressedEnvelope<DnsResponse, InetSocketAddress>> promise;
-    private final int id;
+    private final short id;
     private final DnsQuestion question;
     private final Iterable<DnsRecord> additional;
     private final DnsRecord optResource;
